Add new chat logs for high density demonstration, updating infrastructure to support 'double' time representation.

+++ BotCore/Core/Messaging/ChatMessage.cs
@@ -13,10 +13,10 @@ public class ChatMessage
     public string Username { get; set; }
     public string Message { get; init; }
     public string UserColor { get; init; }
-    public int OffsetSeconds { get; init; }
+    public double OffsetSeconds { get; init; }
     public string Timestamp { get; init; }
 
-    public ChatMessage(string user, string msg, string userColor, int offset = 0, string time = "")
+    public ChatMessage(string user, string msg, string userColor, double offset = 0, string time = "")
     {
         Username = user ?? throw new ArgumentNullException(nameof(user));
         Message = msg ?? throw new ArgumentNullException(nameof(msg));
